% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_of_day_binner.R
\name{aggregate_by_bedtime}
\alias{aggregate_by_bedtime}
\title{Find average circadian fingerprint for each (rounded) bedtime}
\usage{
aggregate_by_bedtime(bins, bedtimes)
}
\arguments{
\item{bins}{a time bins object that lists the circadian footprint of each user}

\item{bedtimes}{a bedtimes object with columns author and nightbed; nightbed is bedtime where 0=midngiht, -12=noon}
}
\value{
circadian meta-footprint for each rounded bedtime
}
\description{
Find average circadian fingerprint for each (rounded) bedtime
}
\examples{
my_timestamps <- data.frame(author=c(1,1,2,2), created_utc=1.6*10^9+c(0,1000,2000,3000))
my_bedtimes <- data.frame(author=c(1,1,2,2), nightbed=c(-1,1))
my_timezones <- create_dummy_timezones(my_timestamps)
my_bins <- time_of_day_binner(assign_time_of_day(my_timestamps, my_timezones))
aggregate_by_bedtime(my_bins, my_bedtimes)
}
