% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_beddit.R
\name{apply_beddit}
\alias{apply_beddit}
\title{Estimate bedtimes of users from circadian fingerpints and model}
\usage{
apply_beddit(bins, model = default_model)
}
\arguments{
\item{bins}{circadian fingerpints of users}

\item{model}{a unified idealized model}
}
\value{
inferred bedtimes of users
}
\description{
Estimate bedtimes of users from circadian fingerpints and model
}
\examples{
apply_beddit(validation_binned_local, default_model)
}
