% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_of_day_binner.R
\name{assign_time_of_day}
\alias{assign_time_of_day}
\title{Add human readable columns to a timestamps object}
\usage{
assign_time_of_day(timestamps, timezones)
}
\arguments{
\item{timestamps}{a dataframe with columns author and created_UTC}

\item{timezones}{a dataframe with columns author and timezone}
}
\value{
a timestamps object that includes extra columns for hour, 15-minute-interval within the hour, date, and day of week
}
\description{
Add human readable columns to a timestamps object
}
\examples{
my_timestamps <- data.frame(author=c(1,1,2,2), created_utc=1.6*10^9+c(0,1000,2000,3000))
my_timezones <- create_dummy_timezones(my_timestamps)
assign_time_of_day(my_timestamps, my_timezones)
}
