% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_beddit.R
\name{bin_beddit}
\alias{bin_beddit}
\title{Convert timestamps data frame into circadian footprint}
\usage{
bin_beddit(timestamps, timezones)
}
\arguments{
\item{timestamps}{a data frame with columns author and created_utc}

\item{timezones}{a data frame with columns author and timezone}
}
\value{
time bins object that gives the circadian footprint of each user
}
\description{
Convert timestamps data frame into circadian footprint
}
\examples{
my_timestamps <- data.frame(author=c(1,1,2,2), created_utc=1.6*10^9+c(0,1000,2000,3000))
bin_beddit(my_timestamps, timezones=FALSE)
}
