% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_of_day_binner.R
\name{clock.mean}
\alias{clock.mean}
\title{Finds the circular mean of a set of clock times}
\usage{
clock.mean(x)
}
\arguments{
\item{x}{clock times (in hours)}
}
\value{
circular mean of the clock times
}
\description{
Finds the circular mean of a set of clock times
}
\examples{
clock.mean(c(-11,11))
}
