% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_bedtimes.R
\name{compared_inferred_to_reported_bedtimes}
\alias{compared_inferred_to_reported_bedtimes}
\title{compare inferred to reported bedtimes}
\usage{
compared_inferred_to_reported_bedtimes(
  reported_bedtimes,
  inferred_bedtimes,
  threshhold = 0.25
)
}
\arguments{
\item{reported_bedtimes}{a data frame where each row contains a username and a reported bedtime}

\item{inferred_bedtimes}{a data frame where each row contains a username and an inferred bedtime}

\item{threshhold}{the minimum correlation between circadian fingerprint and best-matching ideal model, deault 0.25}
}
\value{
a data frame that includes reported_bedtimes and inferred_bedtimes for each user
}
\description{
compare inferred to reported bedtimes
}
\examples{

my_bedtimes <- data.frame(author=c(1,2), nightbed=c(-1, 1))
inferred_bedtimes <- data.frame(author=c(1,2), inferred_bedtime=c(-1.5, 0.75), cor=c(0.5,0.5))
compared_inferred_to_reported_bedtimes(my_bedtimes, inferred_bedtimes)
}
