% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{core_timestamps}
\alias{core_timestamps}
\title{Timestamps of 128 users with public bedtimes and public time zones}
\format{
A data frame with 2,036,843 rows and 2 variables:
\describe{
\item{author}{random identifier for a unique Reddit user}
\item{created_utc}{timestamps, in Unix epoch time format}
...
}
}
\source{
\url{https://psyarxiv.com/9mpbw/}
}
\usage{
core_timestamps
}
\description{
A dataset containing  2,036,843 timestamps of posts of
128 Reddit users with public bedtimes and public time zones.
}
\keyword{datasets}
