% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{core_timezones}
\alias{core_timezones}
\title{Time zones of 128 users with public bedtimes and public time zones}
\format{
A data frame with 128 rows and 2 variables:
\describe{
\item{author}{random identifier for a unique Reddit user}
\item{timezone}{time zone, in Time Zone Database format}
...
}
}
\source{
\url{https://psyarxiv.com/9mpbw/}
}
\usage{
core_timezones
}
\description{
A dataset containing the time zones of 128 Reddit users with publicly implied timezones and
publicly reproted bedtimes.
}
\keyword{datasets}
