% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_of_day_binner.R
\name{counterQ15}
\alias{counterQ15}
\title{Counts the number of instances of values 0 through 95}
\usage{
counterQ15(Z)
}
\arguments{
\item{Z}{integer vector of what 15-minute intervals of the day a set of posts occurred at}
}
\value{
the counts of posts at each 15-minute interval of the day
}
\description{
Counts the number of instances of values 0 through 95
}
\examples{
my_timestamps <- data.frame(author=c(1,1,2,2), created_utc=1.6*10^9+c(0,1000,2000,3000))
my_timezones <- create_dummy_timezones(my_timestamps)
my_timestamps <- assign_time_of_day(my_timestamps, my_timezones)
my_timestamps$q15 <- 4*my_timestamps$hour + 4*my_timestamps$m15
 hx <- my_timestamps[!duplicated(my_timestamps[,c("author", "q15", "date")]),]
 af <- stats::aggregate(q15 ~ author, FUN=counterQ15, hx)
}
