% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_of_day_binner.R
\name{create_dummy_timezones}
\alias{create_dummy_timezones}
\title{Creates a default (UTC) timezones object for a set of timestamps}
\usage{
create_dummy_timezones(timestamps)
}
\arguments{
\item{timestamps}{a dataframe with columns author and created_UTC}
}
\value{
a timezones object: a data frame with 1 row for each author, with timezone listed as "UTC"
}
\description{
Creates a default (UTC) timezones object for a set of timestamps
}
\examples{
my_timestamps <- data.frame(author=c(1,1,2,2), created_utc=1.6*10^9+c(0,1000,2000,3000))
create_dummy_timezones(my_timestamps)
}
