% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_bedtimes.R
\name{estimate_bedtimes}
\alias{estimate_bedtimes}
\title{Estimate bedtimes of users from circadian fingerpints and model}
\usage{
estimate_bedtimes(bins, model)
}
\arguments{
\item{bins}{circadian fingerpints of users}

\item{model}{a unified idealized model}
}
\value{
inferred bedtimes of users
}
\description{
Estimate bedtimes of users from circadian fingerpints and model
}
\examples{
my_timestamps <- data.frame(author=c(1,1,2,2), created_utc=1.6*10^9+c(0,1000,2000,3000))
my_bedtimes <- data.frame(author=c(1,1,2,2), nightbed=c(-1,-1))
my_timezones <- create_dummy_timezones(my_timestamps)
my_bins <- time_of_day_binner(assign_time_of_day(my_timestamps, my_timezones))
rh <- seq(-8,15.75,0.25)
my_model <- data.frame(relhour=rh, idealFreq=(rh/15)^2)
estimate_bedtimes(my_bins, my_model)
}
