% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_bedtimes.R
\name{filter_sufficient}
\alias{filter_sufficient}
\title{filters aggregated by bedtime object to include rows representing at least threshhold authors}
\usage{
filter_sufficient(mag, threshhold = 10)
}
\arguments{
\item{mag}{aggregated-by-bedtime data frame}

\item{threshhold}{the minimum number of authors required to include a bedtime}
}
\value{
filtered aggregated-by-bedtime data frame
}
\description{
filters aggregated by bedtime object to include rows representing at least threshhold authors
}
\examples{
my_timestamps <- data.frame(author=c(1,1,2,2), created_utc=1.6*10^9+c(0,1000,2000,3000))
my_bedtimes <- data.frame(author=c(1,2), nightbed=c(-1,1))
my_timezones <- create_dummy_timezones(my_timestamps)
my_bins <- time_of_day_binner(assign_time_of_day(my_timestamps, my_timezones))
my_binsb <- aggregate_by_bedtime(my_bins, my_bedtimes)
my_mag <- multi_relativizer(my_binsb)
filter_sufficient(my_mag, 1)
}
