% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_bedtimes.R
\name{find_best_parabola}
\alias{find_best_parabola}
\title{Fits unified idealized model}
\usage{
find_best_parabola(mag, out_prefix = "my_prefix", write = F)
}
\arguments{
\item{mag}{bedtime-relativized posting frequencies aggregated by bedtime}

\item{out_prefix}{filename prefix}

\item{write}{set to TRUE if you wish to write the model to disk}
}
\value{
unified idealized model
}
\description{
Fits unified idealized model
}
\examples{
my_timestamps <- data.frame(author=c(1,1,2,2), created_utc=1.6*10^9+c(0,1000,2000,3000))
my_bedtimes <- data.frame(author=c(1,2), nightbed=c(-1,1))
my_timezones <- create_dummy_timezones(my_timestamps)
my_bins <- time_of_day_binner(assign_time_of_day(my_timestamps, my_timezones))
my_binsb <- aggregate_by_bedtime(my_bins, my_bedtimes)
my_mag <- multi_relativizer(my_binsb)
my_mag <- filter_sufficient(my_mag, 1)
find_best_parabola(my_mag)
}
