% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_bedtimes.R
\name{idealXreal_compare}
\alias{idealXreal_compare}
\title{Plots unified model vs observed posting frequencies}
\usage{
idealXreal_compare(obsFreq, idealFreq)
}
\arguments{
\item{obsFreq}{data frame with columns relhour and obsFreq, which gives Reddit posting frequency in observed data}

\item{idealFreq}{data frame with columns relhour and idealFreq, which gives Reddit posting frequency in unified idealized model}
}
\value{
plotting function
}
\description{
Plots unified model vs observed posting frequencies
}
\examples{
rh <- seq(-8,15.75,0.25)
obsFreq <- data.frame(relhour=rh, obsFreq=(rh/15)^2+runif(96)/10)
idealFreq <- data.frame(relhour=rh, idealFreq=(rh/15)^2)
idealXreal_compare(obsFreq, idealFreq)
}
