% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_bedtimes.R
\name{multi_relativizer}
\alias{multi_relativizer}
\title{Relativizes the bins of each bedtime within a bins_bedtime data frame}
\usage{
multi_relativizer(binsb)
}
\arguments{
\item{binsb}{a data frame with cols 3:98 as midnight-relativized freqs and there is bedtime col}
}
\value{
a data frame with cols 3:98 as bedtime-relativized freqs and there is bedtime col
}
\description{
Relativizes the bins of each bedtime within a bins_bedtime data frame
}
\examples{
my_timestamps <- data.frame(author=c(1,1,2,2), created_utc=1.6*10^9+c(0,1000,2000,3000))
my_bedtimes <- data.frame(author=c(1,2), nightbed=c(-1,1))
my_timezones <- create_dummy_timezones(my_timestamps)
my_bins <- time_of_day_binner(assign_time_of_day(my_timestamps, my_timezones))
my_binsb <- aggregate_by_bedtime(my_bins, my_bedtimes)
multi_relativizer(my_binsb)
}
