% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_bedtimes.R
\name{shifter}
\alias{shifter}
\title{cyclically permutes a vector}
\usage{
shifter(x, n = 1)
}
\arguments{
\item{x}{vector to be permuted}

\item{n}{how far to permute}
}
\value{
pertmuted vector
}
\description{
cyclically permutes a vector
}
\examples{
x <- 0:95
shifter(x, 4)
}
