% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_beddit.R
\name{test_beddit}
\alias{test_beddit}
\title{Estimate bedtimes and compare to reported bedtimes}
\usage{
test_beddit(timebins, model, bedtimes, threshold = 0.25)
}
\arguments{
\item{timebins}{circadian fingerpints of users}

\item{model}{an idealized model of circadian posting}

\item{bedtimes}{reported bedtimes of users}

\item{threshold}{how good a fit required to output result}
}
\value{
a data frame containing both reported and inferred bedtimes
}
\description{
Estimate bedtimes and compare to reported bedtimes
}
\examples{
test_beddit(validation_binned_local, default_model, validation_reportedbedtimes)
}
