% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_of_day_binner.R
\name{time_of_day_binner}
\alias{time_of_day_binner}
\title{Creates circadian fingerprint}
\usage{
time_of_day_binner(timestamps)
}
\arguments{
\item{timestamps}{a timestamps object that includes human-readable columns}
}
\value{
time bins object that gives the circadian footprint of each user
}
\description{
Creates circadian fingerprint
}
\examples{
my_timestamps <- data.frame(author=c(1,1,2,2), created_utc=1.6*10^9+c(0,1000,2000,3000))
my_timezones <- create_dummy_timezones(my_timestamps)
time_of_day_binner(assign_time_of_day(my_timestamps, my_timezones))
}
