% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_beddit.R
\name{train_beddit}
\alias{train_beddit}
\title{Train a model from a collection of authors' circadian footprints and their known bedtimes}
\usage{
train_beddit(timebins, bedtimes, necessary_authors = 10)
}
\arguments{
\item{timebins}{circadian footprint of each author in authors' local time}

\item{bedtimes}{data frame with columns author and nightbed where 0=midnight}

\item{necessary_authors}{in training model only include bedtimes represented by at least this many authors}
}
\value{
idealized unified model of Reddit posting frequency at each time bin relative to bedtime
}
\description{
Train a model from a collection of authors' circadian footprints and their known bedtimes
}
\examples{
my_timestamps <- data.frame(author=c(1,1,2,2), created_utc=1.6*10^9+c(0,1000,2000,3000))
my_bedtimes <- data.frame(author=c(1,2), nightbed=c(-1,1))
my_timezones <- create_dummy_timezones(my_timestamps)
my_bins <- time_of_day_binner(assign_time_of_day(my_timestamps, my_timezones))
train_beddit(my_timebins, my_bedtimes)
}
