% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_bedtimes.R
\name{which.maximum}
\alias{which.maximum}
\title{helper function to make which.max unbiased and single-valued}
\usage{
which.maximum(vec)
}
\arguments{
\item{vec}{vector}
}
\value{
index of maximum element of vector; if multiple, select one at random
}
\description{
helper function to make which.max unbiased and single-valued
}
\examples{
x <- rep(1:5,3)
which.maximum(x)
}
