% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_beddit.R
\name{train_beddit}
\alias{train_beddit}
\title{Train a model from a collection of authors' circadian footprints and their known bedtimes}
\usage{
train_beddit(timebins, bedtimes, necessary_authors = 10)
}
\arguments{
\item{timebins}{circadian footprint of each author in authors' local time}

\item{bedtimes}{data frame with columns author and nightbed where 0=midnight}

\item{necessary_authors}{in training model only include bedtimes represented by at least this many authors}
}
\value{
idealized unified model of Reddit posting frequency at each time bin relative to bedtime
}
\description{
Train a model from a collection of authors' circadian footprints and their known bedtimes
}
\examples{
core_bins <- bin_beddit(timestamps = core_timestamps, timezones = core_timezones)
train_beddit(timebins = core_bins, bedtimes = core_reportedbedtimes)
}
